requirements.txt
Django==1.5.1
MySQL-python==1.2.4
argparse==1.2.1
distribute==0.6.38
dj-database-url==0.2.1
django-concurrency==0.5b1
djangorestframework==2.1.17
psycopg2==2.4.6
requests==1.1.0
wsgiref==0.1.2
